feat: split config into correctness & style
This commit separates rules pertaining to "correctness" of code from
rules that enforce good (in the eyes of Proxima) code style. While both
are important, correctness rules should invoke significantly less debate
than style rules. This will allow atypical styled code bases to only
extend the correctness rules and still benefit from improved safety
without having to adopt all of the stylistic rule set.

We also introduce a new style "standard" as opposed to the previous
"verbose". The key difference between standard and verbose is that
standard makes it optional to declare the type of a variable when the
compiler can safely infer its type. This trades of code readability that
comes with the verbose style, for succinctness and safety (it is
possible to sometimes lead the TypeScript compiler astray when declaring
ones own types for every variable). Types are still required on module
boundaries and class members.

Co-authored-by: Oliver Chalk <oliverchalk245@gmail.com>